src/backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asttokens==2.4.1
    # via stack-data
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
bokeh==3.5.2
    # via
    #   panel
    #   umap-learn
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   dask
    #   typer
    #   uvicorn
cloudpickle==3.0.0
    # via dask
colorama==0.4.6
    # via
    #   click
    #   ipython
    #   tqdm
    #   uvicorn
colorcet==3.1.0
    # via
    #   datashader
    #   holoviews
    #   umap-learn
comm==0.2.2
    # via
    #   ipykernel
    #   ipywidgets
contourpy==1.3.0
    # via
    #   bokeh
    #   matplotlib
cycler==0.12.1
    # via matplotlib
dask==2024.8.2
    # via datashader
datashader==0.16.3
    # via umap-learn
debugpy==1.8.5
    # via ipykernel
decorator==5.1.1
    # via ipython
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
executing==2.1.0
    # via stack-data
fastapi==0.112.2
    # via -r requirements.in
fastapi-cli==0.0.5
    # via fastapi
fastjsonschema==2.20.0
    # via nbformat
filelock==3.15.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.53.1
    # via matplotlib
fsspec==2024.6.1
    # via
    #   dask
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
holoviews==1.14.9
    # via umap-learn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via fastapi
huggingface-hub==0.24.6
    # via
    #   tokenizers
    #   transformers
idna==3.8
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
imageio==2.35.1
    # via scikit-image
ipykernel==6.29.5
    # via -r requirements.in
ipython==8.27.0
    # via
    #   ipykernel
    #   ipywidgets
ipywidgets==8.1.5
    # via -r requirements.in
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via
    #   bokeh
    #   fastapi
    #   torch
joblib==1.4.2
    # via
    #   pynndescent
    #   scikit-learn
jsonschema==4.23.0
    # via nbformat
jsonschema-specifications==2023.12.1
    # via jsonschema
jupyter-client==8.6.2
    # via ipykernel
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
    #   nbformat
jupyterlab-widgets==3.0.13
    # via ipywidgets
kiwisolver==1.4.5
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
llvmlite==0.43.0
    # via
    #   numba
    #   pynndescent
locket==1.0.0
    # via partd
markdown==3.7
    # via panel
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.2
    # via
    #   seaborn
    #   umap-learn
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multipledispatch==1.0.0
    # via datashader
nbformat==5.10.4
    # via -r requirements.in
nest-asyncio==1.6.0
    # via ipykernel
networkx==3.3
    # via
    #   scikit-image
    #   torch
numba==0.60.0
    # via
    #   datashader
    #   pynndescent
    #   umap-learn
numpy==1.26.4
    # via
    #   -r requirements.in
    #   bokeh
    #   contourpy
    #   datashader
    #   holoviews
    #   imageio
    #   matplotlib
    #   numba
    #   pandas
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   seaborn
    #   tifffile
    #   transformers
    #   umap-learn
    #   xarray
packaging==24.1
    # via
    #   bokeh
    #   dask
    #   datashader
    #   holoviews
    #   huggingface-hub
    #   ipykernel
    #   lazy-loader
    #   matplotlib
    #   scikit-image
    #   transformers
    #   xarray
pandas==2.2.2
    # via
    #   bokeh
    #   datashader
    #   holoviews
    #   seaborn
    #   umap-learn
    #   xarray
panel==0.10.3
    # via holoviews
param==1.13.0
    # via
    #   datashader
    #   holoviews
    #   panel
    #   pyct
    #   pyviz-comms
parso==0.8.4
    # via jedi
partd==1.4.2
    # via dask
pillow==10.4.0
    # via
    #   bokeh
    #   datashader
    #   imageio
    #   matplotlib
    #   scikit-image
platformdirs==4.2.2
    # via jupyter-core
prompt-toolkit==3.0.47
    # via ipython
psutil==6.0.0
    # via ipykernel
pure-eval==0.2.3
    # via stack-data
pyct==0.5.0
    # via
    #   datashader
    #   panel
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via
    #   ipython
    #   rich
pynndescent==0.5.13
    # via umap-learn
pyparsing==3.1.4
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.9
    # via fastapi
pytz==2024.1
    # via pandas
pyviz-comms==3.0.3
    # via
    #   holoviews
    #   panel
pywin32==306
    # via jupyter-core
pyyaml==6.0.2
    # via
    #   bokeh
    #   dask
    #   huggingface-hub
    #   transformers
    #   uvicorn
pyzmq==26.2.0
    # via
    #   ipykernel
    #   jupyter-client
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.7.24
    # via transformers
requests==2.32.3
    # via
    #   datashader
    #   huggingface-hub
    #   panel
    #   transformers
rich==13.8.0
    # via typer
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
safetensors==0.4.4
    # via transformers
scikit-image==0.24.0
    # via umap-learn
scikit-learn==1.5.1
    # via
    #   -r requirements.in
    #   pynndescent
    #   umap-learn
scipy==1.14.1
    # via
    #   datashader
    #   pynndescent
    #   scikit-image
    #   scikit-learn
    #   umap-learn
seaborn==0.13.2
    # via umap-learn
setuptools==74.0.0
    # via torch
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
stack-data==0.6.3
    # via ipython
starlette==0.38.4
    # via fastapi
sympy==1.13.2
    # via torch
threadpoolctl==3.5.0
    # via scikit-learn
tifffile==2024.8.30
    # via scikit-image
tokenizers==0.19.1
    # via transformers
toolz==0.12.1
    # via
    #   dask
    #   datashader
    #   partd
torch==2.4.0
    # via -r requirements.in
tornado==6.4.1
    # via
    #   bokeh
    #   ipykernel
    #   jupyter-client
tqdm==4.66.5
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   panel
    #   transformers
    #   umap-learn
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbformat
transformers==4.44.2
    # via -r requirements.in
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   typer
tzdata==2024.1
    # via pandas
umap-learn==0.5.6
    # via -r requirements.in
urllib3==2.2.2
    # via requests
uvicorn==0.30.6
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
watchfiles==0.24.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==13.0.1
    # via uvicorn
widgetsnbextension==4.0.13
    # via ipywidgets
xarray==2024.7.0
    # via datashader
xyzservices==2024.9.0
    # via bokeh
